% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agEval.R
\name{agEval}
\alias{agEval}
\title{Aggregate the Performance Matrices of Multiple Interpolations}
\usage{
agEval(pmats, task)
}
\arguments{
\item{pmats}{The performance matrices (result of performance.R)}

\item{task}{Type of aggregation (mean, quantile distribution, etc.)}
}
\description{
Function to 'condense' the performance matrices of the interpolations from each method (M) across K simulations in each (P,G) missingness specification
}
