% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileMatrix.R
\name{compileMatrix}
\alias{compileMatrix}
\title{Compile agEvaluate Objects into Matrices}
\usage{
compileMatrix(agEval)
}
\arguments{
\item{agEval}{A list object (result of agEvaluate.R) of aggregated performance metrics}
}
\description{
A function to compile objects of class "agEvaluate" (from agEvaluate()) into matrices of dimension PxG.\cr
 Elements of agEvaluate are arranged into this matrix corresponding to (p,g).
}
