% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CStable.R
\name{CStable}
\alias{CStable}
\title{Display Performance Criteria Across Method}
\usage{
CStable(d = 1, agEval, m = names(agEval[[1]][[1]][[1]]),
  crit = "MSE", f = "median", cross_section = "g",
  layer_type = "method", fixedIndex = NULL, collapse = T)
}
\arguments{
\item{d}{An integer that is the index of the dataset of interest}

\item{agEval}{A list object (result of agEval.R) of aggregated performance metrics}

\item{m}{A vector of the interpolation methods of interest}

\item{crit}{A character string describing the performance metric of interest}

\item{f}{"mean" or "median" (default); which statistic to use for f(p,g)}

\item{cross_section}{"g" or "p"; specifies which variable to hold constant in the table (gap width or proportion missing)}

\item{fixedIndex}{An integer specifying the index position of the variable in "fixed" to hold constant in the table}

\item{collapse}{logical; Generate a table collapsing across the fixed variable (T), or select a fixed index (over 'fixed') (F)}
}
\description{
Function to display the results of the performance analysis and produce an xtable() for input into LaTeX.
}
