% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_gradient.R
\name{eval_gradient}
\alias{eval_gradient}
\title{Report on the Stability of Interpolation Algorithms}
\usage{
eval_gradient(
  agObject,
  metric,
  m = names(agObject[[1]][[1]][[1]]),
  d = names(agObject),
  cm,
  cd,
  f = "median"
)
}
\arguments{
\item{agObject}{\code{aggregate_pf}; A list object (result of \code{aggregate_pf()}) of aggregated performance metrics}

\item{metric}{\code{character}; A character vector describing the performance metrics of interest}

\item{m}{\code{character}; A vector of the interpolation methods of interest}

\item{d}{\code{character}; A vector of the datasets of interest}

\item{cm}{\code{character}; The chosen method(s) of interest}

\item{cd}{\code{character}; The chosen dataset(s) of interest}

\item{f}{\code{character}; "median" (default): which statistic will be represented by the gradient}
}
\description{
A function that will produce a report on the stability of the input interpolation algorithms according to the gradient.
}
