% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_performance.R
\name{eval_performance}
\alias{eval_performance}
\title{Evaluate Interpolation Performance of a Single Time Series}
\usage{
eval_performance(x, X)
}
\arguments{
\item{x}{The original time series vector}

\item{X}{The interpolated time series vector}
}
\description{
Function to define and store performance criteria for the comparison of a single interpolated series with its original.
Criteria are "optimized" when they are either maximized (1) or minimized (0) as per the following:
# R = 1
R^2 = 1
abs_differences = 0
MBE = 0
ME = 0
MAE = 0
MRE = 0
MARE = 0
MAPE = 0
SSE = 0 
MSE = 0
RMS = 0 
NMSE = 0
RE = 1
RMSE = 0
NRMSD = 0
RMSS = 0
}
