% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaps.R
\name{gaps}
\alias{gaps}
\title{Generate Gaps in a Vector}
\usage{
gaps(x, prop_missing, gap_width)
}
\arguments{
\item{x}{\code{numeric}; The complete time series vector from which to remove values}

\item{prop_missing}{\code{numeric}; Proportion of observations to remove, *P*}

\item{gap_width}{\code{integer}; Length of gap (how many successive observations to remove), *G*}
}
\description{
Subroutine of \code{simulateGaps()}. Function to generate gaps of a specified length constrained by proportion of missingness.
}
\examples{
 
n = 100

# Remove prop_missing*n points where the number of consecutive missing
# observations (gaps) is a multiple of gap_width

gappyx <- gaps(rnorm(n), prop_missing = 0.30, gap_width = 2)

# will generate 30 missing points, with <15 holes

}
