% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient.R
\name{gradient}
\alias{gradient}
\title{Approximate the Gradient}
\usage{
gradient(
  d = 1:length(agEval),
  m = names(agEval[[1]][[1]][[1]]),
  crit,
  agEval,
  f = "median"
)
}
\arguments{
\item{d}{A vector of the indexes of the datasets of interest}

\item{m}{A vector of the interpolation methods of interest}

\item{crit}{A character vector describing the performance metrics of interest}

\item{agEval}{A list object (result of agEval.R) of aggregated performance metrics}

\item{f}{"median" (default): which statistic will be represented by the gradient}
}
\description{
A function that will calculate the slope between every combination of adjacent points in a grid.
}
