% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapGrid.R
\name{heatmapGrid}
\alias{heatmapGrid}
\title{Generate Heatmap Objects}
\usage{
heatmapGrid(agEval, f = "median", crit, m, d, type = "z",
  output = "graphic", col = c("red", "yellow"))
}
\arguments{
\item{agEval}{A list object of class "\code{agEvaluate}" (result of \code{agEvaluate()} of aggregated performance metrics}

\item{f}{"\code{median}" (default): Which statistic should be used that will be represented in the heatmap}

\item{crit}{numeric; a vector of performance criteria}

\item{m}{numeric; a vector describing the index positions of desired methods (corresponding to \code{agEval})}

\item{d}{numeric; a vector describing the index positions of desired datasets (corresponding to \code{agEval})}

\item{type}{"\code{z}" (default) or "\code{gradient}"; Generate heatmaps of the matrix of PxG values of the performance metrics\cr 
in \code{crit} (for \code{type} ='\code{z}'), or the values of the slopes between adjacent values in the performance matrices \cr
(for \code{type} = '\code{gradient}')}

\item{output}{"\code{graphic}" (default) or "\code{matrix}"; If \code{type = "z"}, \code{output} is coerced to "\code{graphic}". Otherwise, if \cr
\code{type = "gradient"} and \code{output = "matrix"}, the function will return the matrix of slopes arranged in a grid corresponding to each\cr
adjacent point in the performance matrix.}

\item{col}{character; A vector of colours to use for the heatmap graphic.}
}
\description{
A function to generate heatmaps to visualize results from objects of class "\code{agEvaluate}" (from \code{agEvaluate()}). \cr
}
