% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapGrid2.R
\name{heatmapGrid2}
\alias{heatmapGrid2}
\title{Generate Heatmaps of Performance}
\usage{
heatmapGrid2(agEval, f = "median", crit, m, d = 1:length(agEval),
  colors = c("#F9E0AA", "#F7C65B", "#FAAF08", "#FA812F", "#FA4032",
  "#F92111"))
}
\arguments{
\item{agEval}{\code{agEvaluate}; An object containing the aggregated performance metrics (result of \code{agEvaluate()})}

\item{f}{\code{character}; The statistic of interest that will be depicted by the heatmap. Possible choices are listed in \code{?agEvaluate}.}

\item{crit}{\code{character}; A single element describing the performance metric of interest}

\item{m}{\code{character}; A single element describing the interpolation methods of interest}

\item{d}{\code{numeric}; A vector to indicate datasets of interest}

\item{colors}{\code{character}; A vector of the desired color palette, with entries in HTML format (\code{"#xxxxxx"})}
}
\description{
A function to generate a set of heatmaps depicting the performances of particular sets of interpolations. This graphic can be considered as a series of 'compressed' surface plots, arranged by dataset.\cr
}
