% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiCS.R
\name{multiCS}
\alias{multiCS}
\title{Arrange Multiple Collapsed Cross-section Plots on a Grid}
\usage{
multiCS(
  agEval,
  d = 1:length(agEval),
  m = names(agEval[[1]][[1]][[1]]),
  crit,
  layer_type = "method",
  f = "median",
  cross_section = "p",
  highlight = "HWI",
  highlight_color = "#EE5C42",
  colors = c("#F9E0AA", "#F7C65B", "#FAAF08", "#FA812F", "#FA4032", "#F92111")
)
}
\arguments{
\item{agEval}{\code{agEvaluate}; An object containing the aggregated performance metrics (result of \code{agEvaluate()})}

\item{d}{\code{numeric}; A vector to indicate datasets of interest}

\item{m}{\code{character}; A vector of interpolation methods of interest}

\item{crit}{\code{character}; An element describing the performance metric of interest}

\item{layer_type}{\code{character}; How to layer the ribbons (by "method" or by "dataset")}

\item{f}{\code{character}; The statistic of interest defining the ribbon. Possible choices are listed in \code{?agEvaluate}}

\item{cross_section}{\code{character}; An element describing the gap structure variable to represent on the x-axis: either \code{"p"} or \code{"g"}}

\item{highlight}{\code{character/numeric}; A single method (if \code{layer_type = "method"}) or dataset (if \code{layer_type = "dataset"}) to highlight.}

\item{highlight_color}{\code{character}; An HTML color of format \code{"#xxxxxx"} to apply to \code{highlight}}

\item{colors}{\code{character}; A vector of the desired color palette, with entries in HTML format (\code{"#xxxxxx"})}
}
\description{
Function to generate a grid of collapsed cross-section plots to faciliate comparison of the 'collapsed' performance metrics of interest
across \strong{method} or \strong{dataset}. Details of each individual plot structure in \code{?plotCS}.
}
