% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiHeatmap.R
\name{multiHeatmap}
\alias{multiHeatmap}
\title{Arrange Multiple Heatmaps on a Grid}
\usage{
multiHeatmap(
  crit,
  agEval,
  m,
  by = "crit",
  f = "median",
  d = 1:length(agEval),
  colors = c("#F9E0AA", "#F7C65B", "#FAAF08", "#FA812F", "#FA4032", "#F92111")
)
}
\arguments{
\item{crit}{\code{character}; A vector describing the performance metrics of interest}

\item{agEval}{\code{agEvaluate}; An object containing the aggregated performance metrics (result of \code{agEvaluate()})}

\item{m}{\code{character}; A vector describing the interpolation methods of interest}

\item{by}{\code{character}; Either \code{"crit"} or \code{"method"}}

\item{f}{\code{character}; The statistic of interest that will be depicted by the heatmap. Possible choices are listed in \code{?agEvaluate}.}

\item{d}{\code{numeric}; A vector to indicate datasets of interest}

\item{colors}{\code{character}; A vector of the desired color palette, with entries in HTML format (\code{"#xxxxxx"})}
}
\description{
Function to generate a grid of heatmaps to compare the performances of particular sets of interpolations across a set of __criteria__ or __methods__.
\itemize{
\item If \code{by = "crit"}, rows of heatmaps for a chosen method are arranged by \strong{criterion}. \cr
\item If \code{by = "method"}, rows of heatmaps for a chosen criterion are arranged by \strong{method}. \cr
}
}
