% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiSurface.R
\name{multiSurface}
\alias{multiSurface}
\title{Arrange Multiple Surface Plots on a Grid}
\usage{
multiSurface(
  d = 1:length(agObject),
  m = names(agObject[[1]][[1]][[1]]),
  metric,
  agObject,
  layer_type = "method",
  f = "median",
  highlight = "HWI",
  highlight_color = "#FA4032",
  colors = c("#F9E0AA", "#F7C65B", "#FAAF08", "#FA812F", "#FA4032", "#F92111")
)
}
\arguments{
\item{d}{\code{numeric}; A vector to indicate datasets of interest}

\item{m}{\code{character}; A vector of interpolation methods of interest}

\item{metric}{\code{character}; An element describing the performance metric of interest}

\item{agObject}{\code{aggregate_pf}; An object containing the aggregated performance metrics (result of \code{aggregate_pf()})}

\item{layer_type}{\code{character}; How to layer the surfaces (by \code{"method"} or by \code{"dataset"})}

\item{f}{\code{character}; The statistic of interest defining the surface \code{f(p,g)}. Possible choices are listed in \code{?aggregate_pf}.}

\item{highlight}{\code{character/numeric}; A single method (if \code{layer_type = "method"}) or dataset (if \code{layer_type = "dataset"}) to highlight.}

\item{highlight_color}{\code{character}; An HTML color of format \code{"#xxxxxx"} to apply to \code{highlight}}

\item{colors}{\code{character}; A vector of the desired color palette, with entries in HTML format (\code{"#xxxxxx"})}
}
\description{
Function to generate a grid of surface plots (from the \code{plotly} package) to visualize changes in the performance metrics of interest
with gap structure, across a number of datasets or methods. The structure of each surface plot is described in \code{?plotSurface}. \cr \cr
\itemize{
\item If \code{layer_type = "method"}, plots are arranged in a grid, by \strong{dataset}. \cr
\item If \code{layer_type = "dataset"}, plots will be arranged in a grid, by \strong{method}. \cr
}
}
