% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parInterpolate.R
\name{parInterpolate}
\alias{parInterpolate}
\title{Parallel Interpolation}
\usage{
parInterpolate(GappyList, methods = NULL, FUN_CALL = NULL,
  numCores = detectCores(), parallel = "k")
}
\arguments{
\item{GappyList}{A list of dimension P x G x K containing gappy time series.}

\item{methods}{vector of IDs for selected interpolation methods, where m = 1,...,M}

\item{FUN_CALL}{User specified interpolation function(s) to be applied to GappyList. Must be a character string in the form: `function_name(args = ..., x = `.}

\item{numCores}{How many CPU cores to use. The default is to use the total number of available cores, as determined by `detectCores()`.}

\item{parallel}{Over which index to parallelize. Possible choices: "p","g","k"}
}
\description{
Function to perform interpolation on gappy series in parallel using user-specified and/or user-defined interpolation methods. Parallelization occurs over the K level. \cr\cr
INTERPOLATION METHOD IDS\cr
1 - Nearest Neighbor - NN\cr
2 - Linear Interpolation - LI\cr
3 - Natural Cubic Spline - NCS\cr
4 - FMM Cubic Spline - FMM\cr 
5 - Hermite Cubic Spline - HCS\cr
6 - Stineman Interpolation - SI\cr
7 - Kalman ARIMA - KAF\cr
8 - Kalman StructTS - KSF\cr
9 - Last Observation Carried Forward - LOCF\cr
10 - Next Observation Carried Backward - NOCB\cr
11 - Simple Moving Average - SMA\cr
12 - Linear Weighted Moving Average - LWMA\cr
13 - Exponential Weighted Moving Average - EWMA\cr
14 - Replace with Mean - RMEA\cr
15 - Replace with Median - REMD\cr
16 - Replace with Mode - RMOD\cr
17 - Replace with Random -RRND\cr
18 - Hybrid Wiener Interpolator - HWI\cr
}
\examples{
# Built-in interpolators
methods <- c(17,5) # Replace with Random, Hermite Cubic Spline

# User-defined functions to pass to FUN_CALL
## Toy function 1: Convert each value of x to its index position
plus <- function(x){
vec <- numeric(length(x))
for(i in 1:length(vec)){
 vec[i] <- i
}
return(vec)
}

## Toy function 2: Convert each value of x to its negative index position
minus <- function(x){
 vec <- numeric(length(x))
 for(i in 1:length(vec)){
   vec[i] <- -i
 }
 return(vec)
}

FUN_CALL <- c("plus(","minus(")

IntData <- list()

for(d in 1:length(OriginalData)){
 IntData[[d]] <- parInterpolate(GappyList = GappyData[[d]], methods = methods, FUN_CALL = FUN_CALL)
}
names(IntData) <- names(OriginalData)
}
