% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAbsDiff.R
\name{plotAbsDiff}
\alias{plotAbsDiff}
\title{Plot Interpolation Error}
\usage{
plotAbsDiff(OriginalData, IntData, d, p, g, m, output = "plot")
}
\arguments{
\item{OriginalData}{A list containing the original time series}

\item{IntData}{A list containing the interpolated time series}

\item{d}{A vector of datasets of interest}

\item{p}{A vector of 'proportion missings' of interest}

\item{g}{A vector of 'gap length' of interest}

\item{m}{A vector of interpolation methods of interest}

\item{output}{A character variable; "plot" outputs a list of plots to be called, "numeric" outputs a list of numeric vectors corresponding to each plot}
}
\description{
Function to generate a nested list of plots to visualize the absolute difference between the original time series and the average (across k simulations) interpolated time series under particular missingness scenarios.
}
