% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCS_un.R
\name{plotCS_un}
\alias{plotCS_un}
\title{Plot Uncollapsed Cross-sections by 'Gap Width' or 'Proportion Missing'}
\usage{
plotCS_un(
  agEval,
  cross_section = "p",
  crit,
  d = 1:length(agEval),
  m = names(agEval[[1]][[1]][[1]]),
  f = "median",
  layer_type = "method",
  highlight = "HWI",
  highlight_color = "#FA4032",
  colors = c("#FF8633", "#FFAF33", "#FFD133", "#FFEC33", "#D7FF33", "#96FF33")
)
}
\arguments{
\item{agEval}{\code{agEvaluate}; An object containing the aggregated performance metrics (result of \code{agEvaluate()})}

\item{cross_section}{\code{character}; An element describing the gap structure variable to represent on the x-axis: either \code{"p"} or \code{"g"}}

\item{crit}{\code{character}; An element describing the performance metric of interest}

\item{d}{\code{numeric}; A vector to indicate datasets of interest}

\item{m}{\code{character}; A vector of interpolation methods of interest}

\item{f}{\code{character}; The statistic of interest defining the ribbon. Possible choices are listed in \code{?agEvaluate}}

\item{layer_type}{\code{character}; How to layer the ribbons (by "method" or by "dataset")}

\item{highlight}{\code{character/numeric}; A single method (if \code{layer_type = "method"}) or dataset (if \code{layer_type = "dataset"}) to highlight.}

\item{highlight_color}{\code{character}; An HTML color of format \code{"#xxxxxx"} to apply to \code{highlight}}

\item{colors}{\code{character}; A vector of the desired color palette, with entries in HTML format (\code{"#xxxxxx"})}
}
\description{
Function to generate uncollapsed cross-section plots to visualize changes in the performance metrics of interest
as the \strong{proportion of missing data} and \strong{gap width increase}. Instead of three axes, as in the surface plots, 
either \code{p} or \code{g} can be broken down such that changes in performance are depicted with respect to one variable, 
across each value of the other. The variable to form the x-axis is specified in \code{cross_section}.
Plots are arranged vertically, where each represents performance plotted against each fixed value in the 
variable not specified in \code{cross_section}. The middle line is the \strong{median} value, the upper 
ribbon boundary is the \strong{97.5th quantile} value, and the lower ribbon boundary is the
\strong{2.5th quantile} value.
}
