% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotError.R
\name{plotError}
\alias{plotError}
\title{Plot Interpolation Error}
\usage{
plotError(
  OriginalData,
  IntData,
  GappyData,
  d,
  p,
  g,
  m,
  output = "plot",
  error_color = "black",
  k = NULL,
  alpha = 0.3
)
}
\arguments{
\item{OriginalData}{\code{list}; A list containing the original time series}

\item{IntData}{\code{list}; A list containing the interpolated time series}

\item{GappyData}{\code{list}; A list containing the gappy time series}

\item{d}{\code{character}; A single character element describing the dataset of interest}

\item{p}{\code{numeric}; A single numeric element describing the of 'proportion missing' of interest}

\item{g}{\code{numeric}; A single numeric element describing the 'gap length' of interest}

\item{m}{\code{character}; A single character element describing the interpolation method of interest}

\item{output}{\code{character}; "plot" outputs a plot object, "numeric" outputs a numeric vector corresponding to the plot}

\item{error_color}{\code{character}; A character element describing the color (HTML: "#XXXXXX") to use for the error ribbon}

\item{k}{\code{numeric}; Numeric element(s) describing the index positions of individual interpolated time series}

\item{alpha}{\code{numeric}; A numeric value describing the opacity of the error ribbon}
}
\description{
Function to generate a plot to visualize the absolute difference between the original time series and the interpolated time series under a specific gap structure.  If more than one value of \code{k} is selected, the resulting plot represents the average interpolation.
}
