% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMetrics.R
\name{plotMetrics}
\alias{plotMetrics}
\title{Plot Grid of Performance Metrics}
\usage{
plotMetrics(agEval, d, p, m, g, metrics = c("MSE", "MAPE",
  "abs_differences"))
}
\arguments{
\item{agEval}{A list containing the aggregated performance metrics (result of agEval.R)}

\item{d}{A vector of datasets of interest (must be a subset of agEval)}

\item{p}{A vector of 'proportion missings' of interest (must be a subset of agEval)}

\item{m}{A vector of interpolation methods of interest (must be a subset of agEval)}

\item{g}{A vector of 'gap length' of interest (must be a subset of agEval)}

\item{metrics}{A character vector of performance metrics of interest}
}
\description{
Function to generate a grid of plots for visualizing a selection of performance metrics.
}
