% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMt.R
\name{simMt}
\alias{simMt}
\title{Simulate M_t}
\usage{
simMt(n = 1000, numTrend = 0, trendType = "polynomial")
}
\arguments{
\item{n}{The length of the output series}

\item{numTrend}{The number of terms to include in the varying trend component of M_t. Value represents the maximum degree; all lower order terms included.}
}
\description{
Function to simulate M_t: the mean component of X_t.
}
