% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateGaps.R
\name{simulateGaps}
\alias{simulateGaps}
\title{Simulate Gappy Data}
\usage{
simulateGaps(data, prop_vec, gap_vec, K)
}
\arguments{
\item{data}{The original time series vector}

\item{prop_vec}{Vector of missingness proportions}

\item{gap_vec}{Vector of gap lengths}

\item{K}{Number of gappy series to simulate for each gap width and proportion missing specification}
}
\description{
Function to simulate the gappy data for a single dataset and store in a list. Function requires gaps.R
}
\examples{
prop_vec = c(0.05,0.10,0.15,0.20)
gap_vec = c(1,5,10)
K = 10 # number of gappy series to simulate under each p,g specification

GappyData <- list()

for(d in 1:length(OriginalData)){
  GappyData[[d]] <- simulateGaps(data = as.numeric(OriginalData[[d]]), prop_vec = prop_vec, gap_vec = gap_vec, K = K)
}
names(GappyData) <- names(OriginalData)
# dimension (d, p, g, k) 
}
