% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitData.R
\name{splitData}
\alias{splitData}
\title{Split Data}
\usage{
splitData(GappyData, split = split)
}
\arguments{
\item{GappyData}{A list object of dimension DxPxGxK}

\item{split}{The number of groups to divide K into (ie. K/split)}
}
\description{
Function to K-split the Gappy Data
}
