% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitchData.R
\name{stitchData}
\alias{stitchData}
\title{Stitch Data}
\usage{
stitchData(directory)
}
\arguments{
\item{directory}{The directory location of the .rda files to stitch. Must be devoid of other files and each file must be identical in structure.}
}
\description{
Function to stitch together data
}
