% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztable.R
\name{ztable}
\alias{ztable}
\title{Create a Table of Values (LaTeX)}
\usage{
ztable(agObject, d, crit, m, sdist = F, f = NULL, LaTeX = T)
}
\arguments{
\item{agObject}{\code{aggregate_pf}; An object containing the aggregated performance metrics (result of \code{aggregate_pf()})}

\item{d}{\code{numeric}; A single value to indicate the dataset of interest}

\item{crit}{\code{character}; A single element describing the performance metric of interest}

\item{m}{\code{character}; A single element describing the interpolation method of interest}

\item{sdist}{\code{logical}; \code{TRUE} returns a table of the sampling distribution \code{(Q2.5, median, Q97.5)} of the chosen metric at every combination of \code{(p,g)}. \code{FALSE} returns a table of values corresponding to \code{f(p,g)}, where each data point is the chosen sample statistic at each \code{(p,g)}.}

\item{f}{\code{character}; If \code{sdist = F}, the sample statistic of interest defining \code{f(p,g)}. Possible choices are listed in \code{?aggregate_pf}.}

\item{LaTeX}{\code{logical}; \code{TRUE} returns a table in LaTeX format. \code{FALSE} returns a table in raw matrix format.}
}
\description{
A function to produce tables of data compatible with LaTeX. \cr
}
