% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agEvaluate.R
\name{agEvaluate}
\alias{agEvaluate}
\title{Aggregate the Performance Matrices of Multiple Interpolations}
\usage{
agEvaluate(pmats)
}
\arguments{
\item{pmats}{\code{list}; A nested list of dimension D x M x P x G x K (result of \code{performance.R}), where the terminal node is a performance matrix.}
}
\description{
Function to aggregate the set of performance matrices, by criterion, using sample statistics of the sampling distribution across K. Resulting object is of class \code{'agEvaluate'}.
}
\details{
The statistics provided in the output are as follows:
\itemize{
\item mean
\item standard deviation 
\item minimum (0\% quantile)
\item 2.5\% quantile
\item 25\% quantile
\item median (50\% quantile)
\item 75\% quantile
\item 97.5\% quantile
\item maximum (100\% quantile)
\item IQR (97.5\% quantile - 2.5\% quantile)
\item skewness
\item p-value of dip test for unimodality (see \code{?dip} for details)
}
}
