% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agEvaluate.R
\name{agEvaluate}
\alias{agEvaluate}
\title{Aggregate the Performance Matrices of Multiple Interpolations}
\usage{
agEvaluate(pmats)
}
\arguments{
\item{pmats}{The performance matrices (result of performance.R)}

\item{task}{Type of aggregation. "default" returns the matrices, "errors" returns the}
}
\description{
Function to 'condense' the performance matrices of the interpolations from each method (M) across K simulations in each (P,G) missingness specification
}
