% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agEvaluate.R
\name{agEvaluate}
\alias{agEvaluate}
\title{Aggregate the Performance Matrices of Multiple Interpolations}
\usage{
agEvaluate(pmats)
}
\arguments{
\item{pmats}{\code{list}; A nested list of dimension D x M x P x G x K (result of \code{performance.R}), where the terminal node is a performance matrix.}
}
\description{
Function to aggregate the set of performance matrices, by criterion, using sample statistics of the sampling distribution across K. Resulting object is of class \code{'agEvaluate'}.
}
\details{
The statistics provided in the output are as follows:
\itemize{
\item \code{(mean)}; mean
\item \code{(sd)}; standard deviation 
\item \code{(q0)}; minimum (0\% quantile)
\item \code{(q2.5)}; 2.5\% quantile
\item \code{(q25)}; 25\% quantile
\item \code{(median)}; median (50\% quantile)
\item \code{(q75)}; 75\% quantile
\item \code{(q97.5)}; 97.5\% quantile
\item \code{(q100)}; maximum (100\% quantile)
\item \code{(iqr)}; IQR (75\% quantile - 25\% quantile)
\item \code{(skewness)}; skewness
\item \code{(dip)}; p-value of dip test for unimodality (see \code{?dip} for details)
}
}
