% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileMatrix.R
\name{compileMatrix}
\alias{compileMatrix}
\title{Compile 'aggregate_pf' Objects into Matrices of Z values}
\usage{
compileMatrix(agObject)
}
\arguments{
\item{agObject}{A list object (result of \code{aggregate_pf()}) of aggregated performance metrics}
}
\description{
A function to compile objects of class \code{"aggregate_pf"} (from \code{aggregate_pf()}) into matrices of dimension PxG.\cr
 Elements of \code{agObject} are arranged into this matrix corresponding to (p,g).
}
