% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileMatrix.R
\name{compileMatrix}
\alias{compileMatrix}
\title{Compile 'aggregate' Objects into Matrices}
\usage{
compileMatrix(agObject)
}
\arguments{
\item{agObject}{A list object (result of \code{aggregate()}) of aggregated performance metrics}
}
\description{
A function to compile objects of class \code{"aggregate"} (from \code{aggregate()}) into matrices of dimension PxG.\cr
 Elements of \code{aggregate} are arranged into this matrix corresponding to (p,g).
}
