% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaps.R
\name{gaps}
\alias{gaps}
\title{Generate Gaps in a Vector}
\usage{
gaps(x, prop_missing, gap_width)
}
\arguments{
\item{x}{The complete vector from which to remove values}

\item{prop_missing}{Proportion of observations to remove, P}

\item{gap_width}{Length of gap (how many successive observations to remove), G}
}
\description{
Function to generate gaps of a specified length constrained by proportion of missingness. Meant to be paired with simulateGaps.R
}
