% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient.R
\name{gradient}
\alias{gradient}
\title{Approximate the Gradient}
\usage{
gradient(
  agObject,
  d = names(agObject),
  m = names(agObject[[1]][[1]][[1]]),
  metric,
  f = "median"
)
}
\arguments{
\item{agObject}{\code{aggregate_pf}; A list object (result of \code{aggregate_pf()}) of aggregated performance metrics}

\item{d}{\code{character}; A vector of the datasets of interest}

\item{m}{\code{character}; A vector of the interpolation methods of interest}

\item{metric}{\code{character}; A character vector describing the performance metrics of interest}

\item{f}{\code{character}; "median" (default): which statistic will be represented by the gradient}
}
\description{
A function that will calculate the slope between every combination of adjacent points in a grid.
}
