% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parInterpolate.R
\name{parInterpolate}
\alias{parInterpolate}
\title{Parallel Interpolation}
\usage{
parInterpolate(gappyTS, methods)
}
\arguments{
\item{gappyTS}{A gappy time series vector}

\item{methods}{vector of IDs for selected interpolation methods, M}
}
\description{
Function to perform interpolation on gappy series in parallel using user-specified interpolation methods.
INTERPOLATION METHOD IDS
1 - Nearest Neighbor
2 - Linear Interpolation
3 - Natural Cubic Spline
4 - FMM Cubic Spline
5 - Hermite Cubic Spline
6 - Stineman Interpolation
7 - Kalman - ARIMA
8 - Kalman - StructTS
9 - Last Observation Carried Forward
10 - Next Observation Carried Backward
11 - Simple Moving Average
12 - Linear Weighted Moving Average
13 - Exponential Weighted Moving Average
14 - Replace with Mean
15 - Replace with Median
16 - Replace with Mode
17 - Replace with Random
18 - Hybrid Wiener
}
