% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCS.R
\name{plotCS}
\alias{plotCS}
\title{Plot Collapsed Cross-sections by 'Gap Width' or 'Proportion Missing'}
\usage{
plotCS(
  d = 1:length(agEval),
  m = names(agEval[[1]][[1]][[1]]),
  crit,
  agEval,
  layer_type = "method",
  f = "median",
  cross_section = "p",
  highlight = NULL,
  highlight_color = "#FF0000",
  colors = c("#FF8633", "#FFAF33", "#FFD133", "#FFEC33", "#D7FF33", "#96FF33")
)
}
\arguments{
\item{d}{\code{numeric}; A vector to indicate datasets of interest}

\item{m}{\code{character}; A vector of interpolation methods of interest}

\item{crit}{\code{character}; An element describing the performance metric of interest}

\item{agEval}{\code{agEvaluate}; An object containing the aggregated performance metrics (result of \code{agEvaluate()})}

\item{layer_type}{\code{character}; How to layer the ribbons (by "method" or by "dataset")}

\item{f}{\code{character}; The statistic of interest defining the ribbon. Possible choices are listed in \code{?agEvaluate}}

\item{cross_section}{\code{character}; An element describing the gap structure variable to represent on the x-axis: either \code{"p"} or \code{"g"}}

\item{highlight}{\code{character/numeric}; A single method (if \code{layer_type = "method"}) or dataset (if \code{layer_type = "dataset"}) to highlight.}

\item{highlight_color}{\code{character}; An HTML color of format \code{"#xxxxxx"} to apply to \code{highlight}}

\item{colors}{\code{character}; A vector of the desired color palette, with entries in HTML format (\code{"#xxxxxx"})}
}
\description{
Function to generate collapsed cross-section plots to visualize changes in the performance metrics of interest
as the \strong{proportion of missing data} and \strong{gap width increase}. Instead of three axes, as in the surface plots, 
either \code{p} or \code{g} can be 'collapsed' such that changes in performance are depicted with respect to one variable, 
across all values of the other. This variable is specified in \code{cross_section}, and will form the x-axis of the cross-section plot.
\itemize{
\item The middle line is the \strong{median} value of the collapsed variable \cr
\item The upper ribbon boundary is the \strong{maximum} value \cr
\item The lower ribbon boundary is the \strong{minimum} value \cr
}
}
