% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMetrics.R
\name{plotMetrics}
\alias{plotMetrics}
\title{Plot Grid of Performance Metrics}
\usage{
plotMetrics(
  agObject,
  d = names(agObject),
  p,
  g,
  m = names(agObject[[1]][[1]][[1]]),
  metric,
  highlight_color = "mistyrose2"
)
}
\arguments{
\item{agObject}{\code{aggregate_pf}; A list containing the aggregated performance metrics (result of \code{aggregate_pf()})}

\item{d}{\code{character}; A vector of datasets of interest}

\item{p}{\code{numeric}; A vector of 'proportion missings' of interest}

\item{g}{\code{numeric}; A vector of 'gap length' of interest}

\item{m}{\code{character}; A vector of interpolation methods of interest}

\item{metric}{\code{character}; A character vector of performance metrics of interest}
}
\description{
Function to generate a grid of plots for visualizing a selection of performance metrics (median).
}
