% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMetrics.R
\name{plotMetrics}
\alias{plotMetrics}
\title{Plot Grid of Performance Metrics}
\usage{
plotMetrics(
  agEval,
  d = 1:length(agEval),
  p = 1:length(agEval[[1]]),
  m = 1:length(agEval[[1]][[1]][[1]]),
  g = 1:length(agEval[[1]][[1]]),
  crit
)
}
\arguments{
\item{agEval}{A list containing the aggregated performance metrics (result of agEval.R)}

\item{d}{A vector of datasets of interest (must be a subset of agEval)}

\item{p}{A vector of 'proportion missings' of interest (must be a subset of agEval)}

\item{m}{A vector of interpolation methods of interest (must be a subset of agEval)}

\item{g}{A vector of 'gap length' of interest (must be a subset of agEval)}

\item{crit}{A character vector of performance metrics of interest}
}
\description{
Function to generate a grid of plots for visualizing a selection of performance metrics (median).
}
