% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSkew.R
\name{plotSkew}
\alias{plotSkew}
\title{Aggregate and plot the skewness statistics over all simulations}
\usage{
plotSkew(
  agObject,
  plotEach = TRUE,
  show_symmetric = NULL,
  output = "plot",
  metric = rownames(agObject[[1]][[1]][[1]][[1]])
)
}
\arguments{
\item{agObject}{\code{aggregate_pf}; A list object (result of \code{aggregate_pf()}) of aggregated performance metrics}

\item{plotEach}{\code{logical}; Whether to display plots individually (histograms) or together as a density plot}

\item{output}{\code{character}; "plot" (default) or "table".}

\item{symmetric}{\code{logical}; TRUE = Display only symmetric metrics, FALSE = Display only asymmetric metrics, NULL = Display all metrics}
}
\description{
Function to display the sampling distributions of the skewness statistics for each metric across all simulations. 
Sampling distributions with a sample mean roughly centered at zero indicate that on average the metric has a reliably 
symmetric distribution.  Distributions of skewness values whose center is not in the neighbourhood of zero (-s/3 <= Xbar <= s/3) indicate that 
on average the metric has a skewed distribution, and the median will better capture the essence of the data.
}
