% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSurface.R
\name{plotSurface}
\alias{plotSurface}
\title{Plot Surface}
\usage{
plotSurface(d = 1:length(agEval), m = names(agEval[[1]][[1]][[1]]),
  crit, agEval, layer_type = "method", f = "median",
  highlight = "HWI", highlight_colour = "#EE5C42")
}
\arguments{
\item{d}{A vector of the indexes of the datasets of interest}

\item{m}{character; A vector of the interpolation methods of interest (maximum of 5)}

\item{crit}{A character vector describing the performance metrics of interest}

\item{agEval}{A list object (result of agEval.R) of aggregated performance metrics}

\item{layer_type}{"method" (default) or "dataset"; how to slice the data}

\item{f}{"median" (default); which statistic to use for f(p,g)}
}
\description{
Function to generate a surface plot to visualize changes in the performance metrics of interest
as the proportion of missing data and gap width increase. Metrics are aggregated across K simulations in each p,g,d,m specification. \cr
The x-axis represents p, the proportion of missing data.\cr 
The y-axis represents g, the gap width.\cr
The z-axis represents the value of the performance metric of interest.
}
