% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotXt.R
\name{plotXt}
\alias{plotXt}
\title{Plot Each Component of Xt}
\usage{
plotXt(d)
}
\arguments{
\item{d}{The index of the original time series of interest}
}
\description{
A function to generate a triptych to visualize each component of the original time series
}
