% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMt.R
\name{simMt}
\alias{simMt}
\title{Simulate M_t}
\usage{
simMt(N = 1000, numTrend = 0, a = NULL, mu = NULL, center = NULL)
}
\arguments{
\item{N}{\code{integer}; The length of the output series}

\item{numTrend}{\code{integer}; The number of terms to include in the varying trend component of \eqn{M_t}. Value represents the maximum degree; all lower order terms included.}

\item{a}{\code{numeric}; If specified, a vector of custom coefficients for each term in the polynomial.}

\item{mu}{\code{numeric}; If specified, a single value representing the grand mean of the polynomial.}

\item{center}{\code{numeric}; If specified, a vector of custom centering parameters for each term in the polynomial.}
}
\description{
Subroutine of \code{simXt()}. Function to simulate \eqn{M_t}: the mean component of \eqn{X_t}.
}
