% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMt.R
\name{simMt}
\alias{simMt}
\title{Simulate M_t}
\usage{
simMt(N = 1000, numTrend = 0, trendType = "polynomial")
}
\arguments{
\item{N}{The length of the output series}

\item{numTrend}{The number of terms to include in the varying trend component of M_t. Value represents the maximum degree; all lower order terms included.}

\item{trendType}{The type of trends to include. Either "polynomial" or "exponential".}
}
\description{
Function to simulate M_t: the mean component of X_t.
}
