% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simTt.R
\name{simTt}
\alias{simTt}
\title{Simulate T_t}
\usage{
simTt(n = 1000, numFreq = 20, bandwidth = NULL)
}
\arguments{
\item{n}{The length of the output series}

\item{numFreq}{The number of sinusoids to generate}

\item{bandwidth}{Value is plugged into a negative exponent (base 10) and defines some interval in which to sample frequencies. If specified, bandwidth must be at least -log10(1/numFreq). If unspecified, freq =  runif(numFreq,0,1)}
}
\description{
Function to simulate T_t: the trend component of X_t.
}
