% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateGaps.R
\name{simulateGaps}
\alias{simulateGaps}
\title{Simulate Gappy Data}
\usage{
simulateGaps(OriginalData, prop_vec, gap_vec, K)
}
\arguments{
\item{OriginalData}{\code{list}; List object containing the complete original time series vectors (as separate elements)}

\item{prop_vec}{\code{numeric}; Vector of unique proportions representing the different 'proportion-of-data missing' (*p*) scenarios to simulate}

\item{gap_vec}{\code{integer}; Vector of unique integer values representing the different 'gap length' scenarios (*g*) to simulate}

\item{K}{\code{integer}; Number of gappy series to simulate for each (*p,g*) gap specification}
}
\description{
Function to simulate the MCAR gappy data for a list of complete original time series according to various combinations of gap structure parameters *p* and *g*. and store in a list.
}
\examples{
 

prop_vec = c(0.05,0.10,0.15,0.20)
gap_vec = c(1,5,10)
K = 10 # number of gappy series to simulate under each p,g specification

GappyData <- simulateGaps(OriginalData = OriginalData, prop_vec = prop_vec, gap_vec = gap_vec, K = K)

# dimension (d, p, g, k) 
}
