% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simWt.R
\name{simWt}
\alias{simWt}
\title{Simulate W_t}
\usage{
simWt(n = 1000, p = 0, q = 0, var = 1)
}
\arguments{
\item{n}{The length of the output series}

\item{p}{The AR order}

\item{q}{The MA order}

\item{var}{The variance of the stochastic process}
}
\description{
Function to simulate W_t: the noise component of X_t. Defaults to white noise.
}
